webapp/requirements.txt
flask==2.3.3
torch==2.0.1
torchvision==0.15.2
Pillow==10.0.0
numpy==1.24.3
werkzeug==2.3.7
requests==2.31.0
huggingface_hub==0.17.3
